Remove: socket.chat.js
